requirements.txt
webapp2
GoogleAppEngineCloudStorageClient
Jinja2
